Fix #2

+++ Basic 1/20.cpp
@@ -2,12 +2,12 @@
 using namespace std;
 
 int main() {
-    int digit = 0;
+    uint64_t digit = 0;
     string n;
     cin >> n;
     for (int i = 0; i < n.length(); ++i) {
         digit += n[i] - '0';
     }
-    cout << (stoi(n) % digit == 0);
+    cout << (stol(n) % digit == 0);
     return 0;
 }
